refactor(score-configs): extract dialog form content
